OSIS-5221 move tree factory into own set of file
