+++ program_management/tests/ddd/factories/program_tree.py
@@ -30,7 +30,6 @@ import factory.fuzzy
 
 from base.models.authorized_relationship import AuthorizedRelationshipList
 from base.models.enums.education_group_types import GroupType, TrainingType
-from base.models.enums.education_group_types import MiniTrainingType
 from program_management.ddd.domain.program_tree import ProgramTree
 from program_management.models.enums.node_type import NodeType
 from program_management.tests.ddd.factories.authorized_relationship import AuthorizedRelationshipObjectFactory
@@ -88,226 +87,6 @@ class ProgramTreeFactory(factory.Factory):
         tree_standard.root_node.children = [link1, link2, link3]
 
         return tree_standard
-
-    @staticmethod
-    def produce_standard_2m_tree(current_year: int, end_year: int) -> 'ProgramTree':
-        tree_data = {
-            "node_type": TrainingType.PGRM_MASTER_120,
-            "year": current_year,
-            "end_year": end_year,
-            "node_id": 1,
-            "children": [
-                {
-                    "node_type": GroupType.COMMON_CORE,
-                    "year": current_year,
-                    "end_year": end_year,
-                    "node_id": 21,
-                    "children": [
-                        {
-                            "node_type": GroupType.SUB_GROUP,
-                            "year": current_year,
-                            "end_year": end_year,
-                            "node_id": 31,
-                            "children": [
-                                {
-                                    "node_type": NodeType.LEARNING_UNIT,
-                                    "year": current_year,
-                                    "end_date": end_year,
-                                    "node_id": 41,
-                                },
-                                {
-                                    "node_type": NodeType.LEARNING_UNIT,
-                                    "year": current_year,
-                                    "end_date": end_year,
-                                    "node_id": 42,
-                                }
-                            ]
-                        },
-                        {
-                            "node_type": NodeType.LEARNING_UNIT,
-                            "year": current_year,
-                            "end_date": end_year,
-                            "node_id": 32,
-                        }
-                    ]
-                },
-                {
-                    "node_type": GroupType.FINALITY_120_LIST_CHOICE,
-                    "year": current_year,
-                    "end_year": end_year,
-                    "node_id": 22,
-                    "children": [
-                        {
-                            "node_type": TrainingType.MASTER_MD_120,
-                            "year": current_year,
-                            "end_year": end_year,
-                            "node_id": 33,
-                            "children": [
-                                {
-                                    "node_type": GroupType.COMMON_CORE,
-                                    "year": current_year,
-                                    "end_year": end_year,
-                                    "node_id": 43,
-                                    "children": [
-                                        {
-                                            "node_type": GroupType.SUB_GROUP,
-                                            "year": current_year,
-                                            "end_year": end_year,
-                                            "node_id": 51,
-                                            "children": [
-                                                {
-                                                    "node_type": NodeType.LEARNING_UNIT,
-                                                    "year": current_year,
-                                                    "end_date": end_year,
-                                                    "node_id": 61,
-                                                },
-                                            ]
-                                        },
-                                    ]
-                                },
-                                {
-                                    "node_type": GroupType.OPTION_LIST_CHOICE,
-                                    "year": current_year,
-                                    "end_year": end_year,
-                                    "node_id": 44,
-                                    "children": [
-                                        {
-                                            "node_type": MiniTrainingType.OPTION,
-                                            "year": current_year,
-                                            "end_year": end_year,
-                                            "code": "LOPTION2112",
-                                            "node_id": 52,
-                                        },
-                                    ]
-                                }
-                            ]
-                        }
-                    ]
-                },
-                {
-                    "node_type": GroupType.OPTION_LIST_CHOICE,
-                    "year": current_year,
-                    "end_year": end_year,
-                    "node_id": 23,
-                    "children": [
-                        {
-                            "node_type": MiniTrainingType.OPTION,
-                            "year": current_year,
-                            "end_year": end_year,
-                            "code": "LOPTION2112",
-                            "node_id": 52,
-                        },
-                    ]
-                },
-                {
-                    "node_type": GroupType.COMPLEMENTARY_MODULE,
-                    "year": current_year,
-                    "end_year": end_year,
-                    "node_id": 24,
-                    "children": [
-                        {
-                            "node_type": NodeType.LEARNING_UNIT,
-                            "year": current_year,
-                            "end_date": end_year,
-                            "node_id": 34,
-                        },
-                        {
-                            "node_type": NodeType.LEARNING_UNIT,
-                            "year": current_year,
-                            "end_date": end_year,
-                            "node_id": 35,
-                        },
-                    ]
-                },
-            ]
-
-        }
-        return tree_builder(tree_data)
-
-    @staticmethod
-    def produce_bachelor_tree(current_year: int, end_year: int) -> 'ProgramTree':
-        tree_data = {
-            "node_type": TrainingType.BACHELOR,
-            "year": current_year,
-            "end_year": end_year,
-            "node_id": 1,
-            "children": [
-                {
-                    "node_type": GroupType.COMMON_CORE,
-                    "year": current_year,
-                    "end_year": end_year,
-                    "node_id": 21,
-                    "children": [
-                        {
-                            "node_type": GroupType.SUB_GROUP,
-                            "year": current_year,
-                            "end_year": end_year,
-                            "node_id": 31,
-                            "children": [
-                                {
-                                    "node_type": NodeType.LEARNING_UNIT,
-                                    "year": current_year,
-                                    "end_date": end_year,
-                                    "node_id": 41,
-                                },
-                                {
-                                    "node_type": NodeType.LEARNING_UNIT,
-                                    "year": current_year,
-                                    "end_date": end_year,
-                                    "node_id": 42,
-                                }
-                            ]
-                        }
-                    ]
-                },
-                {
-                    "node_type": GroupType.MINOR_LIST_CHOICE,
-                    "year": current_year,
-                    "end_year": end_year,
-                    "node_id": 22,
-                    "children": [
-                        {
-                            "node_type": MiniTrainingType.DEEPENING,
-                            "year": current_year,
-                            "end_year": end_year,
-                            "node_id": 32,
-                            "children": [
-                                {
-                                    "node_type": GroupType.COMMON_CORE,
-                                    "year": current_year,
-                                    "end_year": end_year,
-                                    "node_id": 1211,
-                                    "children": [
-                                        {
-                                            "node_type": GroupType.SUB_GROUP,
-                                            "year": current_year,
-                                            "end_year": end_year,
-                                            "node_id": 43,
-                                            "children": [
-                                                {
-                                                    "node_type": NodeType.LEARNING_UNIT,
-                                                    "year": current_year,
-                                                    "end_date": end_year,
-                                                    "node_id": 51,
-                                                },
-                                                {
-                                                    "node_type": NodeType.LEARNING_UNIT,
-                                                    "year": current_year,
-                                                    "end_date": end_year,
-                                                    "node_id": 52,
-                                                }
-                                            ]
-                                        }
-                                    ]
-                                }
-                            ]
-                        },
-                    ]
-                },
-                {"node_type": NodeType.LEARNING_UNIT}
-            ]
-        }
-        return tree_builder(tree_data)
 
 
 def _tree_builder(data: Dict) -> 'Node':
